2019 Day 08 and 09
Merged from old 2019 only repo. These commits were made before these repos were merged together, but forgot to push them to main

+++ 2019/Libraries/IntCodeInterpreter/Input/FileInputParser.cs
@@ -9,12 +9,12 @@ namespace IntCodeInterpreter.Input
     {
         #region Instance Methods
 
-        public async Task<List<int>> ReadOperationsFromFile(string fileName)
+        public async Task<List<long>> ReadOperationsFromFile(string fileName)
         {
             var fileText = await File.ReadAllTextAsync(fileName);
 
             var ints = fileText.Split(',')
-                               .Select(int.Parse)
+                               .Select(long.Parse)
                                .ToList();
 
             return ints;
